requirements.txt
matplotlib==3.8.0
numpy==1.26.1
Pillow==10.0.1
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1